Move mock_user_info to test helpers
